--- a/src/DynamoCoreWpf/Controls/InPorts.xaml
+++ b/src/DynamoCoreWpf/Controls/InPorts.xaml
@@ -1,0 +1,9 @@
+<UserControl x:Class="Dynamo.UI.Controls.InPorts"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:p="clr-namespace:Dynamo.Wpf.Properties"
+             xmlns:ui="clr-namespace:Dynamo.UI"
+             mc:Ignorable="d">
+</UserControl>