--- a/src/DynamoCoreWpf/Controls/InPorts.xaml.cs
+++ b/src/DynamoCoreWpf/Controls/InPorts.xaml.cs
@@ -256,6 +256,9 @@ namespace Dynamo.UI.Controls
                 CommandParameter = viewModel
             };
 
+            mouseRightButtonDownTrigger.Actions.Add(mouseRightButtonDownAction);
+            Dynamo.Microsoft.Xaml.Behaviors.Interaction.GetTriggers(MainGrid).Add(mouseRightButtonDownTrigger);
+
             var previewMouseLeftDownTrigger = new Dynamo.UI.Views.HandlingEventTrigger()
             {
                 EventName = "PreviewMouseLeftButtonDown",