--- a/src/DynamoCoreWpf/Controls/OutPorts.xaml
+++ b/src/DynamoCoreWpf/Controls/OutPorts.xaml
@@ -1,0 +1,8 @@
+<UserControl x:Class="Dynamo.UI.Controls.OutPorts"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Dynamo.Wpf.Controls"
+             mc:Ignorable="d">
+</UserControl>