… src/DynamoCoreWpf/DynamoCoreWpf.csproj
… src/DynamoCoreWpf/PublicAPI.Unshipped.txt
… src/DynamoCoreWpf/UI/Themes/Modern/DynamoColorsAndBrushes.xaml
… src/DynamoCoreWpf/UI/Themes/Modern/InPortsOld.xaml
… src/DynamoCoreWpf/UI/Themes/Modern/OutPortsOld.xaml
… src/DynamoCoreWpf/ViewModels/Core/PortViewModel.cs
… src/DynamoCoreWpf/Views/Core/AnnotationView.xaml.cs
… src/DynamoCoreWpf/Views/Core/NodeView.xaml
… src/DynamoCoreWpf/Views/Core/NodeView.xaml.cs
… test/DynamoCoreWpf3Tests/NodeViewCustomizationTests.cs
… test/DynamoCoreWpf3Tests/NodeViewTests.cs