### Purpose

The Purpose of this PR is to convert the Dynamo NodeView Control and InPorts.xaml and OutPorts.xaml to a C# only set of WPF Controls.  The reason is to avoid the overhead of the parsing the xaml in cases where the graph has a large number of nodes.  The code only option also allows for some simplification of the NodeView and PortView by allowing the NodeViews to be mutated in situations where their Properties vary (ie Output Port for CodeBlock vs Standard Nodes) in if statements versus via embed triggers in the xaml.  

This mechanism will vary in terms of performance but generally speaking this PR is 2x to 3x faster in Load time and 50% reduction in RAM memory allocation.

Some Profiling results:
|| Old  |  New|
|----|----|-----|
512 |27s|11s|
Mars |43s|28s|
Manekin |146s|93s|

In sample graph with 512 PointByCoordinate Nodes the load time reduces from ~24sec to ~10sec and the total application RAM allocation after opening the file reduces from 1.2gig to 650mb.

### Declarations

Check these if you believe they are true

- [x] The codebase is in a better state after this PR
- [ ] Is documented according to the [standards](https://github.com/DynamoDS/Dynamo/wiki/Coding-Standards)
- [ ] The level of testing this PR includes is appropriate
- [ ] User facing strings, if any, are extracted into `*.resx` files
- [ ] All tests pass using the self-service CI.
- [ ] Snapshot of UI changes, if any.
- [ ] Changes to the API follow [Semantic Versioning](https://github.com/DynamoDS/Dynamo/wiki/Dynamo-Versions) and are documented in the [API Changes](https://github.com/DynamoDS/Dynamo/wiki/API-Changes) document.
- [ ] This PR modifies some build requirements and the readme is updated
- [ ] This PR contains no files larger than 50 MB

### Release Notes

- Ports and NodeView implemented in code behind

### Reviewers

TBD

### FYIs

@achintyabhat @zeusongit @QilongTang