@zeusongit 2025-05-28
Addressed comments, fixes the white rectangle appearing for right click on ports, reference colors and brushes from SharedDictionary.