@zeusongit 2025-06-05
> @zeusongit marked tests as failing? how many are failing

3 are failing: https://master-5.jenkins.autodesk.com/job/Dynamo/job/DynamoSelfServe/job/pullRequestValidation/17914/
Triggered a new build: https://master-5.jenkins.autodesk.com/job/Dynamo/job/DynamoSelfServe/job/pullRequestValidation/17952/
Will mark them before merge